# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "AtualizarRoteiroRequest": {
        "properties": {
          "unidade_ids": {
            "items": {
              "type": "string"
            },
            "title": "Unidade Ids",
            "type": "array"
          }
        },
        "title": "AtualizarRoteiroRequest",
        "type": "object"
      },
...
      "CriarTipoProcessoRequest": {
        "properties": {
          "nome": {
            "minLength": 1,
            "title": "Nome",
            "type": "string"
          },
          "unidade_ids": {
            "items": {
              "type": "string"
            },
            "title": "Unidade Ids",
            "type": "array"
          }
        },
        "required": [
          "nome"
        ],
        "title": "CriarTipoProcessoRequest",
        "type": "object"
      },
...
      "EtapaRoteiroResponse": {
        "properties": {
          "ordem": {
            "title": "Ordem",
            "type": "integer"
          },
          "unidade_id": {
            "title": "Unidade Id",
            "type": "string"
          }
        },
        "required": [
          "unidade_id",
          "ordem"
        ],
        "title": "EtapaRoteiroResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RoteiroResponse": {
        "properties": {
          "etapas": {
            "items": {
              "$ref": "#/components/schemas/EtapaRoteiroResponse"
            },
            "title": "Etapas",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "vigente": {
            "title": "Vigente",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "vigente",
          "etapas"
        ],
        "title": "RoteiroResponse",
        "type": "object"
      },
...
      "TipoProcessoResponse": {
        "properties": {
          "ativo": {
            "title": "Ativo",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "roteiro": {
            "$ref": "#/components/schemas/RoteiroResponse"
          }
        },
        "required": [
          "id",
          "nome",
          "ativo",
          "roteiro"
        ],
        "title": "TipoProcessoResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SETES.DOCS API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tipos-processo": {
      "get": {
        "description": "Catálogo de tipos de processo com o roteiro vigente de cada um.",
        "operationId": "listar_tipos_processo_tipos_processo_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TipoProcessoResponse"
                  },
                  "title": "Response Listar Tipos Processo Tipos Processo Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar Tipos Processo",
        "tags": [
          "tipos-processo"
        ]
      },
      "post": {
        "operationId": "criar_tipo_processo_tipos_processo_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarTipoProcessoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TipoProcessoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Criar Tipo Processo",
        "tags": [
          "tipos-processo"
        ]
      }
    },
    "/tipos-processo/{tipo_processo_id}/roteiro": {
      "put": {
        "description": "US 8.2 Cen.2 — cria nova versão vigente; a anterior fica congelada\n(`vigente=false`) e suas `roteiro_etapa` nunca são alteradas — processos já\ncriados sob ela mantêm o roteiro do momento de sua criação.",
        "operationId": "atualizar_roteiro_tipos_processo__tipo_processo_id__roteiro_put",
        "parameters": [
          {
            "in": "path",
            "name": "tipo_processo_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tipo Processo Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarRoteiroRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoteiroResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualizar Roteiro",
        "tags": [
          "tipos-processo"
        ]
      }
    },
...
  }
}